Many changes (details in description)
* Thanks to @andreyrmg for help with the BMP (issue #4 fix);
* Destroying bullets in a collision with a white border;
* Changing constants;
* Change version of the standard to C++11;
* Variable names fix;
* New map for collision test.

+++ src/Player.h
@@ -1,10 +1,12 @@
 #ifndef __PLAYER__
 #define __PLAYER__
 
+#include "MainConst.h"
+
 class Player{
 public:
 	Player();
-	Player(double _x, double _y, int a);	
+	Player(double _x, double _y, double a);	
 	void drawPlayer();
 	void update();
